# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ledger-as-a-Service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/ledger/layers/{layer_id}/resolve": {
      "post": {
        "summary": "Resolve Review",
        "operationId": "resolve_review_v1_ledger_layers__layer_id__resolve_post",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Layer or review item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Invalid key, decision, or resolution rules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/layers/{layer_id}/initialize": {
      "post": {
        "summary": "Post Initialize Layer",
        "operationId": "post_initialize_layer_v1_ledger_layers__layer_id__initialize_post",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Shell"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Post Initialize Layer V1 Ledger Layers  Layer Id  Initialize Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ledger/layers/{layer_id}/verify": {
      "get": {
        "summary": "Verify Layer Endpoint",
        "operationId": "verify_layer_endpoint_v1_ledger_layers__layer_id__verify_get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "check_signatures",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Check Signatures"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Layer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/layers/{layer_id}/findings": {
      "get": {
        "summary": "Get Findings",
        "operationId": "get_findings_v1_ledger_layers__layer_id__findings_get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Layer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/layers/{layer_id}/cumulative": {
      "get": {
        "summary": "Alias for /findings \u2014 baseline enrichment is a client concern",
        "operationId": "get_cumulative_v1_ledger_layers__layer_id__cumulative_get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Layer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/layers/{layer_id}/events": {
      "get": {
        "summary": "Get Layer Events",
        "operationId": "get_layer_events_v1_ledger_layers__layer_id__events_get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "finding_ref",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a specific finding",
              "title": "Finding Ref"
            },
            "description": "Filter to a specific finding"
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source type",
              "title": "Source Type"
            },
            "description": "Filter by source type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Max events per page",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max events per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Events to skip before returning results",
              "default": 0,
              "title": "Offset"
            },
            "description": "Events to skip before returning results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Layer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/findings": {
      "get": {
        "summary": "Get All Findings",
        "operationId": "get_all_findings_v1_ledger_findings_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resume after this layer_id",
              "title": "Cursor"
            },
            "description": "Resume after this layer_id"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Max layers per page",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max layers per page"
          },
          {
            "name": "since_epoch",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only layers with merkle_epoch >= this value",
              "title": "Since Epoch"
            },
            "description": "Only layers with merkle_epoch >= this value"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkFindingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/layers/{layer_id}/submit": {
      "post": {
        "summary": "Post Submit",
        "operationId": "post_submit_v1_ledger_layers__layer_id__submit_post",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or identity rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/fingerprint": {
      "post": {
        "summary": "Post Fingerprint",
        "operationId": "post_fingerprint_v1_ledger_fingerprint_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FingerprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FingerprintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/layers": {
      "get": {
        "summary": "List Layers",
        "operationId": "list_layers_v1_ledger_layers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ledger/layers/{layer_id}/sign": {
      "post": {
        "summary": "Sign Layer Endpoint",
        "operationId": "sign_layer_endpoint_v1_ledger_layers__layer_id__sign_post",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "rekor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Upload signature to transparency log",
              "default": false,
              "title": "Rekor"
            },
            "description": "Upload signature to transparency log"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sign Layer Endpoint V1 Ledger Layers  Layer Id  Sign Post"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Layer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "500": {
            "description": "Signing failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ledger/whoami": {
      "get": {
        "summary": "Return the token-verified actor for the caller",
        "operationId": "whoami_endpoint_v1_ledger_whoami_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayerActor"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail"
      },
...
      "EventsResponse": {
        "properties": {
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "layer_id": {
            "type": "string",
            "title": "Layer Id"
          }
        },
        "type": "object",
        "required": [
          "events",
          "total",
          "layer_id"
        ],
        "title": "EventsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveRequest": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "key",
          "decision"
        ],
        "title": "ResolveRequest"
      },
      "ResolveResponse": {
        "properties": {
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          },
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "resolved",
          "key"
        ],
        "title": "ResolveResponse"
      },
...
      "VerifyResponse": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/VerifyFinding"
            },
            "type": "array",
            "title": "Findings",
            "default": []
          },
          "checked_at": {
            "type": "string",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "checked_at"
        ],
        "title": "VerifyResponse"
      }
    }
  }
}